+++ tests/testthat/test-groups.R
@@ -49,13 +49,6 @@ test_that("group_model returns correct type", {
   )
 })
 
-test_that("mixture Markov model statistics can be obtained", {
-  expect_error(
-    mmm_stats(engagement_mmm),
-    NA
-  )
-})
-
 test_that("groups can be renamed", {
   model <- group_model(mock_sequence, group = c(1, 1, 2, 2))
   expect_error(
@@ -127,6 +120,14 @@ test_that("groupwise models can be constructed with scaling", {
 test_that("group model can be constructed from a mixture Markov model", {
   expect_error(
     group_model(engagement_tna_mmm),
+    NA
+  )
+})
+
+test_that("group model can be constructed from a clustering result", {
+  clust <- cluster_sequences(engagement, k = 3)
+  expect_error(
+    group_model(clust),
     NA
   )
 })